pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.13",
    "audioop-lts>=0.2.1",
    "discord-py>=2.4.0",
    "mistralai>=1.4.0",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
]
